# 开发技术文档/02-API接口定义文档/openapi.yaml
openapi: "3.0.3"
info:
  title: 知微 ziwi SaaS 平台 API
  description: |
    知微(SaaS)平台 — Phase 1 核心业务 API 文档

    ## 认证方式
    - **Bearer JWT**: `Authorization: Bearer <token>` — 用户身份认证
    - **X-Tenant-Id**: 请求头 — 多租户识别（SaaS 模式必传）
    - **X-Deploy-Mode**: 请求头 — 部署模式: saas / onprem / multi-factory
    - **X-Factory-Id**: 请求头 — 工厂ID（multi-factory 模式必传）
    - **X-License-Key**: 请求头 — 许可证密钥（onprem 模式必传）

    ## API 版本
    - URL 路径版本: `/api/v1/{resource}`
    - 同一接口保留 2 个版本，v1 废弃前 3 个月发送 `X-API-Deprecated` 头

    ## 分页规范
    - `page`: 从 1 开始
    - `page_size`: 默认 20，最大 100
    - `total`: 总记录数
    - 响应格式: `{ "items": [], "total": 0, "page": 1, "page_size": 20 }`

    ## 错误响应格式
    ```json
    {
      "code": "400-1001",
      "message": "参数校验失败：必填字段缺失",
      "request_id": "req_abc123"
    }
    ```

    ## 错误码清单
    | code | 说明 |
    |------|------|
    | 400-0000 | 请求参数错误 |
    | 400-1001 | 参数校验失败：必填字段缺失 |
    | 400-1002 | 参数校验失败：格式错误 |
    | 400-1003 | 参数校验失败：值超出允许范围 |
    | 400-2001 | 业务规则冲突：库存不足 |
    | 400-2002 | 业务规则冲突：状态不允许此操作 |
    | 400-2003 | 业务规则冲突：数据已存在（重复） |
    | 401-0000 | 未认证 |
    | 401-1001 | Token 无效或已过期 |
    | 401-1002 | Token 格式错误 |
    | 401-1003 | OAuth 授权失败 |
    | 403-0000 | 无权限 |
    | 403-1001 | 模块未购买（租户未启用该模块） |
    | 403-1002 | 用户角色无此操作权限 |
    | 403-1003 | 数据范围权限不足（不可跨工厂操作） |
    | 404-0000 | 资源不存在 |
    | 404-1001 | 查询的资源 ID 不存在 |
    | 404-1002 | 关联的父资源不存在 |
    | 409-0000 | 数据冲突 |
    | 409-1001 | 并发修改冲突（乐观锁版本号不一致） |
...
    | 429-0000 | 请求过频繁 |
...
servers:
...
paths:
...
  /api/v1/dictionary:
    get:
      tags: ["M00-字典"]
      summary: 字典列表
      operationId: listDictionaries
      security:
        - BearerAuth: []
        - TenantId: []
    post:
      tags: ["M00-字典"]
      summary: 创建字典
      operationId: createDictionary
      security:
        - BearerAuth: []
        - TenantId: []
  /api/v1/dictionary/{dict_code}/items:
    get:
      tags: ["M00-字典"]
      summary: 字典项列表
      operationId: listDictItems
      security:
        - BearerAuth: []
        - TenantId: []
    post:
      tags: ["M00-字典"]
      summary: 添加字典项
      operationId: createDictItem
      security:
        - BearerAuth: []
        - TenantId: []

  # ==========================================
  # M00 — 团队与员工
  # ==========================================
...
  /api/v1/work-orders:
    get:
      tags: ["M01-工单"]
      summary: 工单列表
      operationId: listWorkOrders
      security:
        - BearerAuth: []
        - TenantId: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: assignee_id
          in: query
          schema:
            type: integer
        - name: due_date_from
          in: query
          schema:
            type: string
            format: date
        - name: due_date_to
          in: query
          schema:
            type: string
            format: date
        - name: keyword
          in: query
          schema:
            type: string
        - name: factory_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: 工单列表
    post:
      tags: ["M01-工单"]
      summary: 创建工单
      operationId: createWorkOrder
      security:
        - BearerAuth: []
        - TenantId: []
  /api/v1/work-orders/{order_id}:
    get:
      tags: ["M01-工单"]
      summary: 工单详情
      operationId: getWorkOrder
      security:
        - BearerAuth: []
        - TenantId: []
    put:
      tags: ["M01-工单"]
      summary: 更新工单
      operationId: updateWorkOrder
      security:
        - BearerAuth: []
        - TenantId: []
...
  /api/v1/work-orders/{order_id}/status-log:
    get:
      tags: ["M01-工单"]
      summary: 工单状态变更日志
      operationId: getWorkOrderStatusLog
      security:
        - BearerAuth: []
        - TenantId: []
...
  /api/v1/work-reports:
    get:
      tags: ["M01-报工"]
      summary: 报工记录列表
      operationId: listWorkReports
      security:
        - BearerAuth: []
        - TenantId: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: work_order_id
          in: query
          schema:
            type: integer
        - name: user_id
          in: query
          schema:
            type: integer
        - name: report_date_from
          in: query
          schema:
            type: string
            format: date
        - name: report_date_to
          in: query
          schema:
            type: string
            format: date
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: 报工列表
    post:
      tags: ["M01-报工"]
      summary: 提交报工
      operationId: createWorkReport
      security:
        - BearerAuth: []
        - TenantId: []
  /api/v1/work-reports/{report_id}:
    get:
      tags: ["M01-报工"]
      summary: 报工详情
      operationId: getWorkReport
      security:
        - BearerAuth: []
        - TenantId: []
    put:
      tags: ["M01-报工"]
      summary: 更新报工
      operationId: updateWorkReport
      security:
        - BearerAuth: []
        - TenantId: []
...
  /api/v1/energy/devices:
    get:
      tags: ["M11-能碳"]
      summary: 能源设备列表
      operationId: listEnergyDevices
      security:
        - BearerAuth: []
        - TenantId: []
...
  /api/v1/energy/alerts:
    get:
      tags: ["M11-能碳"]
      summary: 能耗告警列表
      operationId: listEnergyAlerts
      security:
        - BearerAuth: []
        - TenantId: []
    post:
      tags: ["M11-能碳"]
      summary: 创建能耗告警规则
      operationId: createEnergyAlert
      security:
        - BearerAuth: []
        - TenantId: []

  # ==========================================
  # M12 — IoT 设备
  # ==========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Bearer Token，从登录接口获取"
    TenantId:
      type: apiKey
      in: header
      name: X-Tenant-Id
      description: "租户ID（SaaS模式必传）"
    DeployMode:
      type: apiKey
      in: header
      name: X-Deploy-Mode
      description: "部署模式: saas / onprem / multi-factory"
    FactoryId:
      type: apiKey
      in: header
      name: X-Factory-Id
      description: "工厂ID（multi-factory模式必传）"
    LicenseKey:
      type: apiKey
      in: header
      name: X-License-Key
      description: "许可证密钥（onprem模式必传）"
...
  parameters:
    PageParam:
      name: page
      in: query
      description: 页码，从 1 开始
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSizeParam:
      name: page_size
      in: query
      description: 每页条数，默认 20，最大 100
      schema:
        type: integer
        default: 20
        maximum: 100
        minimum: 1
...